mfd: Add support for X-Powers AC200 EPHY syscon
The X-Powers AC200 mixed signal chip contains a 100Mbit/s Ethernet PHY.
While the PHY is using the standard MDIO and MII/RMII interfaces to
the MAC, there are some registers in the AC200 that need to be setup to
make the PHY functional:
- The MDIO PHY address needs to set.
- The LED polarity needs to be configured.
- The MII interface mode needs to be set (MII or RMII).
- There is a reset line that controls the PHY operation.
- There is a clock gate that blocks or forwards the AC200's internal clock
  to the PHY.

This driver here takes care of those setup needs, but does not cover the
actual PHY operation. Once the PHY is set up and enabled, it behaves like
a standard MDIO/MII controlled PHY, and can be driven by the IEEE802.3-C22
driver.

To some degree those parameters mimic the typical wired setup of a
physical PHY chip: the LED polarity, MII interface mode and PHY address
are typically configued via connecting certain pins. We use the
devicetree to learn those parameters, which depend on the board setup.

This driver is a child of the AC200 MFD parent device, and uses its
regmap and input clock. It also provides a reset and clock controller,
which the actual PHY device (node) needs to refer to. This ensures that
this PHY control device is initialised before the PHY is expected to work.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>