arm64: dts: sun50i-h6-orangepi.dtsi: Rollback r_rsb to r_i2c
This fix affects two boards:
sun50i-h6-orangepi-lite2.dts
sun50i-h6-orangepi-one-plus.dts

It also depends on the revision of the board.
If your board fails when working with this fix, just disable
it by adding (-) minus to the first position of the line
in which the patch is written in the series.conf file.